Rearrange code a bit
